refactor: Time 추가 기능 리팩토링
