+++ src/main/java/roomescape/dto/request/TimeRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 public record TimeRequestDto(
         String time
