+++ src/main/java/roomescape/service/TimeService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.dao.TimeDao;
 import roomescape.domain.Time;
+import roomescape.dto.request.TimeRequestDto;
 
 @Service
 @RequiredArgsConstructor
@@ -15,6 +16,16 @@ public class TimeService {
     @Transactional
     public Time findById(Long timeId) {
         return timeDao.findById(timeId);
+    }
+
+    @Transactional
+    public Time addTime(TimeRequestDto request) {
+        Time time = Time.builder()
+                .time(request.time())
+                .build();
+
+        timeDao.insert(time);
+        return time;
     }
 
 }
